Add the _blurb subdir.
Its purpose is to hold content that will published somewhere else.

Its initial contents are the social media posts from launch, but other
transmittable content should be

The subdir does not strictly need to be situated in the website repo.

However, most of its items will be similar to items to published on the website,
and having it here will simplify any automated pre-processing need for
translation, and simplify the instructions given to translators once we need to
start regularly performing translations in many different languages.

Saving it here also makes the content the valid jekyll collection. If for any
reason part of this content does need to be displayed on the website, there is a
simple route to doing so.